requirements.txt
discord.py[voice]
flask
flask_sqlalchemy
flask_login
python-dotenv
requests
aiohttp
Pillow
matplotlib
psycopg2-binary